trader/requirements.txt
dnspython
pika
requests
pymongo
websocket-client
flask
redis